api_sunat/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.08.22=h06a4308_0
click=8.0.4=py311h06a4308_0
flask=2.2.2=py311h06a4308_0
itsdangerous=2.0.1=pyhd3eb1b0_0
jinja2=3.1.2=py311h06a4308_0
krb5=1.20.1=h143b758_1
ld_impl_linux-64=2.38=h1181459_1
libedit=3.1.20221030=h5eee18b_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libpq=12.15=hdbd6064_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
markupsafe=2.1.1=py311h5eee18b_0
ncurses=6.4=h6a678d5_0
openssl=3.0.11=h7f8727e_2
pip=23.2.1=py311h06a4308_0
psycopg2=2.9.3=py311h5eee18b_1
python=3.11.5=h955ad1f_0
python-decouple=3.8=pypi_0
readline=8.2=h5eee18b_0
setuptools=68.0.0=py311h06a4308_0
sqlite=3.41.2=h5eee18b_0
tk=8.6.12=h1ccaba5_0
tzdata=2023c=h04d1e81_0
werkzeug=2.2.3=py311h06a4308_0
wheel=0.38.4=py311h06a4308_0
xz=5.4.2=h5eee18b_0
zlib=1.2.13=h5eee18b_0
